Repackage endpoints and schema classes (#2)
