+++ fastapi/app/model/request.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel
+
+class Url(BaseModel):
+    address: str
+
+class Comparator(BaseModel):
+    source: str
+    target: str
